// web/projects/portal/src/app/portal/data/model/datasources/database/database-tree-node-type.ts
/*
 * inetsoft-web - StyleBI is a business intelligence web application.
 * Copyright © 2024 InetSoft Technology (info@inetsoft.com)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
export enum DatabaseTreeNodeType {
   DATABASE = "database",
   FOLDER = "folder",
   TABLE_TYPES = "table_types",
   SCHEMA = "schema",
   TABLE = "table",
   COLUMN = "column",
   ALIAS_TABLE_FOLDER = "alias_table_folder",
   PHYSICAL_MODEL = "physical_model",
   PHYSICAL_VIEW = "physical_view",
   ALIAS_TABLE = "alias_table"
}